requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0

# Deep learning
torch>=2.1.0
torchaudio>=2.1.0

# TTS (Coqui XTTS v2 — maintained community fork)
coqui-tts>=0.27.4

# STT (ESPnet OWSM v4)
espnet
espnet_model_zoo

# Audio I/O
soundfile
librosa

# Hugging Face (model download + authentication)
huggingface_hub

# Phoneme / language support
gruut
gruut-lang-en
phonemizer

# Japanese dictionary (required by Coqui TTS internals)
mecab-python3
unidic-lite

# Utilities
numpy
scipy
pyyaml
